[Issue-4263] add: handle for blockstream api
